+++ SensateIoT.Platform.Network.Common/Settings/InternalMqttClientSettings.cs
@@ -1,7 +1,0 @@
-﻿namespace SensateIoT.Platform.Network.Common.Settings
-{
-	public class InternalMqttClientSettings
-	{
-
-	}
-}