+++ SensateIoT.Platform.Network.Data/Abstract/ControlMessageType.cs
@@ -5,8 +5,11 @@
  * @email  michel@michelmegens.net
  */
 
+using JetBrains.Annotations;
+
 namespace SensateIoT.Platform.Network.Data.Abstract
 {
+	[PublicAPI]
 	public enum ControlMessageType
 	{
 		Mqtt = 0,